loaded program
Neural network that is trained with dataset loaded from https://www.kaggle.com/c/digit-recognizer
For program to process image and predict digit, image should be stored in 'images' folder.
Original dataset is a collection of png grayscale images in res 28X28.
Network accuracy is ~97% on validation dataset, but it is lower on actual input (I drew some images in paint to test)
Full project with all dependencies is stored in google disk
https://drive.google.com/open?id=1Ry7ocYvEXTOZEfB-2doQd0pixBV9N5ZC